requirements.txt
attrs==19.1.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Django==1.11.20
django-appconf==1.0.3
django-leaflet==0.24.0
djangorestframework==3.9.1
Fiona==1.8.6
geojson==2.4.1
geopandas==0.5.0
munch==2.3.2
numpy==1.16.1
pandas==0.24.1
psycopg2==2.7.7
pyproj==2.1.3
python-dateutil==2.8.0
pytz==2018.9
rcssmin==1.0.6
rjsmin==1.0.12
Shapely==1.6.4.post2
six==1.12.0
